RFID Folder
You may attempt using this folder instead of doing your own installation of the mercury-api. Simply place this folder on your "desktop" on your PI. You should be able to then follow the instructions in .../RFID/mercuryapi-1.31.5.19/java/README.txt ... that is to say, you will find the instructions there on how to use the code.

I have also added those instructions to the readme here in the RFID branch for your convenience.